requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
ecdsa==0.18.0
streamlit==1.28.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==41.0.7

# Data handling
pandas==2.1.3
numpy==1.26.2

# Visualization (for Streamlit)
plotly==5.18.0

# HTTP requests (for simulator)
requests==2.31.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1